requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==6.2.1
PyYAML==5.3.1
six==1.15.0
tqdm==4.48.2
Werkzeug==1.0.1
waitress==1.4.4
